+++ extensions/iconify/src/components/IconActions.tsx
@@ -1,4 +1,15 @@
-import { Action, ActionPanel, Clipboard, getPreferenceValues, Icon, showToast, Toast } from "@raycast/api";
+import {
+  Action,
+  ActionPanel,
+  Clipboard,
+  getPreferenceValues,
+  Icon,
+  showToast,
+  Toast,
+  LaunchType,
+  open,
+} from "@raycast/api";
+import { crossLaunchCommand } from "raycast-cross-extension";
 import { copyToClipboard, toURL } from "../utils";
 import { iconFormatsMap } from "../utils/format-name";
 import { primaryActionEnum, Preferences } from "../types";
@@ -11,9 +22,10 @@ type IconActionProps = PropsWithChildren<{
   setId?: string;
   svgIcon: string;
   dataURIIcon: string;
+  from?: string;
 }>;
 
-export const IconActions = ({ id, setId, dataURIIcon, svgIcon, children }: IconActionProps) => {
+export const IconActions = ({ id, setId, dataURIIcon, svgIcon, from, children }: IconActionProps) => {
   const formattedIconName = setId ? iconFormatsMap[iconNameFormat]({ setId, iconId: id }) : id;
   const paste = <Action.Paste title="Paste SVG String" content={svgIcon} />;
   const copy = <Action.CopyToClipboard title="Copy SVG String" content={svgIcon} />;
@@ -54,6 +66,39 @@ export const IconActions = ({ id, setId, dataURIIcon, svgIcon, children }: IconA
   const copyName = setId ? <Action.CopyToClipboard title="Copy Name" content={formattedIconName} /> : null;
   const copyURL = setId ? <Action.CopyToClipboard title="Copy URL" content={toURL(setId, id)} /> : null;
   const copyDataURI = <Action.CopyToClipboard title="Copy Data URI" content={dataURIIcon} />;
+
+  function ToolsActionSection() {
+    const pickColor = (
+      <Action
+        title="Pick Color"
+        icon={Icon.EyeDropper}
+        onAction={async () => {
+          try {
+            await crossLaunchCommand({
+              name: "pick-color",
+              type: LaunchType.UserInitiated,
+              extensionName: "color-picker",
+              ownerOrAuthorName: "thomas",
+              context: {
+                copyToClipboard: false,
+                callbackLaunchOptions: from
+                  ? {
+                      name: from,
+                      type: LaunchType.Background,
+                    }
+                  : undefined,
+              },
+            });
+          } catch {
+            open("raycast://extensions/thomas/color-picker");
+          }
+        }}
+      />
+    );
+
+    return <ActionPanel.Section title="Tools">{pickColor}</ActionPanel.Section>;
+  }
+
   return (
     <ActionPanel>
       {primaryAction === primaryActionEnum.paste && (
@@ -152,6 +197,7 @@ export const IconActions = ({ id, setId, dataURIIcon, svgIcon, children }: IconA
           {copyURL}
         </>
       )}
+      <ToolsActionSection />
       {children}
     </ActionPanel>
   );