+++ extensions/umami/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Umami Changelog
 
+## [Windows Support] - 2025-12-19
+
+- Support Windows
+- More precise error handling
+
 ## [Initial Version] - 2025-12-18